Clean up removed range replica on ChangeReplicas commit trigger.
With the addition of the replica being removed to the trigger protobuf
message, as well as a manual add method on the queue, replicas are now
GC'd quickly in the common case. This allows far less pressure to be
put on the range gc queue, as replicas only need be checked against
the definitive range metadata records infrequently.